refactor: make the codec more applicable
