+++ packages/experiment-codec/assets/suggest-trigger.gif
