+++ packages/experiment-codec/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ckb-lumos/experiment-molecule",
+  "name": "@ckb-lumos/experiment-codec",
   "private": true,
   "version": "0.17.0-rc8",
   "description": "Make your own molecule binding in JavaScript(TypeScript)",
@@ -23,14 +23,15 @@
     "lint": "eslint -c ../../.eslintrc.js \"{src,tests,examples}/**/*.ts\"",
     "test": "ava **/*.test.ts --timeout=2m",
     "build": "npm run build:types && npm run build:js",
-    "build:types": "tsc --declaration --emitDeclarationOnly",
+    "build:types": "tsc --declaration --declarationMap --emitDeclarationOnly",
     "build:js": "babel --root-mode upward src --out-dir lib --extensions .ts -s",
     "clean": "rm -rf lib",
     "prepublishOnly": "yarn run clean && yarn run build",
     "release": "npm publish"
   },
   "dependencies": {
-    "@ckb-lumos/bi": "^0.17.0-rc8"
+    "@ckb-lumos/bi": "^0.17.0-rc8",
+    "@ethersproject/bytes": "^5.6.0"
   },
   "publishConfig": {
     "access": "public"
