+++ packages/experiment-codec/src/molecule/index.ts
@@ -1,0 +1,10 @@
+export { byteOf, byteArrayOf, byteVecOf } from "./helper";
+export { table, array, option, struct, vector, union } from "./layout";
+export {
+  UnusedOpt,
+  Bytes,
+  Byte32,
+  Byte32Vec,
+  BytesVec,
+  BytesOpt,
+} from "./common";
