+++ packages/experiment-codec/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "lib"
+    "outDir": "lib",
+    "sourceMap": true
   },
   "include": ["src"]
 }
