--- packages/experiment-molecule/src/index.ts
@@ -1,37 +1,0 @@
-export * as layout from "./layout";
-export * as common from "./common";
-export * as base from "./base";
-
-export {
-  table,
-  option,
-  struct,
-  array,
-  isFixedCodec,
-  union,
-  vector,
-} from "./layout";
-
-export {
-  Uint8,
-  Uint16,
-  Uint32,
-  Uint64,
-  Uint128,
-  Uint256,
-  Uint512,
-  Bytes,
-  BytesOpt,
-  Byte32,
-  UnusedOpt,
-} from "./common";
-
-export { byteOf, byteArrayOf, byteVecOf } from "./base";
-
-export type {
-  BinaryCodec,
-  FixedBinaryCodec,
-  Codec,
-  Pack,
-  Unpack,
-} from "./base";
